Rename to getInitialState
