+++ src/__test__/components/data-exploration/differential-expression-tool/DiffExprManager.test.jsx
@@ -10,7 +10,7 @@ import DiffExprCompute from 'components/data-exploration/differential-expression
 import DiffExprResults from 'components/data-exploration/differential-expression-tool/DiffExprResults';
 
 import initialState from 'redux/reducers/differentialExpression/initialState';
-import constructInitialState from 'redux/reducers/genes/constructInitialState';
+import getInitialState from 'redux/reducers/genes/getInitialState';
 import cellSetsInitialState from 'redux/reducers/cellSets/initialState';
 
 import { mockCellSets } from '__test__/test-utils/cellSets.mock';
@@ -25,7 +25,7 @@ const emptyState = {
     properties: {},
   },
   genes: {
-    ...constructInitialState(),
+    ...getInitialState(),
     focused: undefined,
   },
 };
