+++ src/__test__/components/plots/ViolinPlot.test.jsx
@@ -11,7 +11,7 @@ import {
   initialPlotConfigStates,
 } from 'redux/reducers/componentConfig/initialState';
 import initialExperimentState from 'redux/reducers/experimentSettings/initialState';
-import constructInitialState from 'redux/reducers/genes/constructInitialState';
+import getInitialState from 'redux/reducers/genes/getInitialState';
 import ViolinPlot from 'components/plots/ViolinPlot';
 import '__test__/test-utils/setupTests';
 
@@ -37,7 +37,7 @@ const defaultStore = {
   experimentSettings: {
     ...initialExperimentState,
   },
-  genes: constructInitialState(),
+  genes: getInitialState(),
   backendStatus: {
     [experimentId]: {
       status: {},
