+++ src/__test__/redux/actions/genes/changeGeneSelection.test.js
@@ -1,7 +1,7 @@
 import configureStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
 import changeGeneSelection from 'redux/actions/genes/changeGeneSelection';
-import constructInitialState from 'redux/reducers/genes/constructInitialState';
+import getInitialState from 'redux/reducers/genes/getInitialState';
 
 import { GENES_SELECT, GENES_DESELECT } from 'redux/actionTypes/genes';
 
@@ -12,7 +12,7 @@ const genes = ['a', 'b', 'c', 'd'];
 
 describe('changeGeneSelection action', () => {
   it('Dispatches select event when select event specified', async () => {
-    const store = mockStore(constructInitialState());
+    const store = mockStore(getInitialState());
     await store.dispatch(changeGeneSelection(experimentId, genes, 'select'));
 
     const firstAction = store.getActions()[0];
@@ -21,7 +21,7 @@ describe('changeGeneSelection action', () => {
   });
 
   it('Dispatches deselect event when select event specified', async () => {
-    const store = mockStore(constructInitialState());
+    const store = mockStore(getInitialState());
     await store.dispatch(changeGeneSelection(experimentId, genes, 'deselect'));
 
     const firstAction = store.getActions()[0];
@@ -30,7 +30,7 @@ describe('changeGeneSelection action', () => {
   });
 
   it('Does not dispatch on other choice', async () => {
-    const store = mockStore(constructInitialState());
+    const store = mockStore(getInitialState());
 
     const t = async () => {
       await store.dispatch(changeGeneSelection(experimentId, genes, 'maybeselect'));
