+++ src/__test__/redux/actions/genes/loadPaginatedGeneProperties.test.js
@@ -1,7 +1,7 @@
 import configureStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
 import loadPaginatedGeneProperties from 'redux/actions/genes/loadPaginatedGeneProperties';
-import constructInitialState from 'redux/reducers/genes/constructInitialState';
+import getInitialState from 'redux/reducers/genes/getInitialState';
 
 import { dispatchWorkRequest, seekFromS3 } from 'utils/work/seekWorkResponse';
 
@@ -49,7 +49,7 @@ describe('loadPaginatedGeneProperties action', () => {
   });
 
   it('Does not dispatch when some of the properties are already loading', async () => {
-    const initialState = constructInitialState();
+    const initialState = getInitialState();
     const store = mockStore({
       backendStatus,
       genes:
@@ -67,7 +67,7 @@ describe('loadPaginatedGeneProperties action', () => {
   });
 
   it('Dispatches appropriately on success condition', async () => {
-    const initialState = constructInitialState();
+    const initialState = getInitialState();
     const store = mockStore({
       genes:
       {
@@ -129,7 +129,7 @@ describe('loadPaginatedGeneProperties action', () => {
     const store = mockStore({
       genes:
       {
-        ...constructInitialState(),
+        ...getInitialState(),
       },
       backendStatus,
       networkResources: {
