+++ src/__test__/redux/reducers/genes/markerGenesLoaded.test.js
@@ -1,4 +1,4 @@
-import constructInitialState from 'redux/reducers/genes/constructInitialState';
+import getInitialState from 'redux/reducers/genes/getInitialState';
 import markerGenesLoadedReducer from 'redux/reducers/genes/markerGenesLoaded';
 import { getOneGeneMatrix } from '__test__/utils/ExpressionMatrix/testMatrixes';
 
@@ -6,7 +6,7 @@ describe('markerGenesLoaded', () => {
   it('returns correct state', () => {
     const genes = ['geneA'];
 
-    const newState = markerGenesLoadedReducer(constructInitialState(), {
+    const newState = markerGenesLoadedReducer(getInitialState(), {
       payload: {
         genes,
         data: getOneGeneMatrix('geneA'),
