+++ src/redux/reducers/genes/genesExpressionLoaded.js
@@ -1,6 +1,6 @@
 import _ from 'lodash';
 
-import { initialViewState } from './constructInitialState';
+import { initialViewState } from './getInitialState';
 
 const upperCaseArray = (array) => (array?.map((element) => element.toUpperCase()));
