+++ src/redux/reducers/genes/genesExpressionLoading.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import { initialViewState } from './constructInitialState';
+import { initialViewState } from './getInitialState';
 
 const genesExpressionLoading = (state, action) => {
   const { genes, componentUuid } = action.payload;
