+++ src/redux/reducers/genes/genesPropertiesLoadedPaginated.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import { initialViewState } from './constructInitialState';
+import { initialViewState } from './getInitialState';
 
 const genesPropertiesLoadedPaginated = (state, action) => {
   const {
