+++ src/redux/reducers/genes/markerGenesError.js
@@ -1,11 +1,11 @@
 /* eslint-disable no-param-reassign */
 import produce from 'immer';
 
-import constructInitialState from './constructInitialState';
+import getInitialState from './getInitialState';
 
 const markerGenesError = produce((draft) => {
   draft.markers.loading = false;
   draft.markers.error = true;
-}, constructInitialState());
+}, getInitialState());
 
 export default markerGenesError;
