+++ src/redux/reducers/genes/markerGenesLoading.js
@@ -1,11 +1,11 @@
 /* eslint-disable no-param-reassign */
 import produce from 'immer';
 
-import constructInitialState from './constructInitialState';
+import getInitialState from './getInitialState';
 
 const markerGenesLoading = produce((draft) => {
   draft.markers.loading = true;
   draft.markers.error = false;
-}, constructInitialState());
+}, getInitialState());
 
 export default markerGenesLoading;
